assistant: add Hermes migration surface and MCP
