+++ Dockerfile
@@ -27,12 +27,13 @@ RUN mkdir -p src/bin \
     && echo "fn main() {}" > src/bin/desktop_mcp.rs \
     && echo "fn main() {}" > src/bin/workspace_mcp.rs \
     && echo "fn main() {}" > src/bin/orchestrator_mcp.rs \
+    && echo "fn main() {}" > src/bin/assistant_mcp.rs \
     && cargo build --release --lib 2>/dev/null || true \
     && cargo build --release 2>/dev/null || true
 
 # Copy real source and build
 COPY src/ src/
-RUN cargo build --release --bin sandboxed-sh --bin desktop-mcp --bin workspace-mcp --bin orchestrator-mcp
+RUN cargo build --release --bin sandboxed-sh --bin desktop-mcp --bin workspace-mcp --bin orchestrator-mcp --bin assistant-mcp
 
 # ---------------------------------------------------------------------------
 # Stage 2: Dashboard builder
@@ -97,6 +98,7 @@ COPY --from=rust-builder /build/target/release/sandboxed-sh /usr/local/bin/sandb
 COPY --from=rust-builder /build/target/release/desktop-mcp /usr/local/bin/desktop-mcp
 COPY --from=rust-builder /build/target/release/workspace-mcp /usr/local/bin/workspace-mcp
 COPY --from=rust-builder /build/target/release/orchestrator-mcp /usr/local/bin/orchestrator-mcp
+COPY --from=rust-builder /build/target/release/assistant-mcp /usr/local/bin/assistant-mcp
 
 # -- Copy dashboard standalone build ------------------------------------------
 COPY --from=dashboard-builder /build/dashboard/.next/standalone /opt/dashboard
