+++ dashboard/src/app/assistant/page.tsx
@@ -1,0 +1,3 @@
+import AssistantPage from '@/components/assistant/assistant-page';
+
+export default AssistantPage;
